tutorial_blog_implementation/adk_interactions_integration/pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "adk_interactions_integration"
version = "0.1.0"
description = "ADK Integration with Gemini Interactions API - combining ADK agent patterns with Interactions backend"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.18.0",
    "google-genai>=1.55.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["adk_interactions_agent*"]
